refactor: form 경로 단축
