+++ src/app/(causw)/board/[boardId]/[postId]/[formId]/page.tsx
@@ -4,14 +4,10 @@ import { notFound, useParams } from 'next/navigation';
 
 import { FormProvider, useForm } from 'react-hook-form';
 
-import { ExpiredForm, NoPermissionForm, ResponseForm } from '@/widgets/form/ui';
-
-import { useGetFormResponseInfo } from '@/entities/form/model';
-import { useHandleApplySubmit } from '@/entities/form/model/hooks';
-import { useResponseFormStore } from '@/entities/form/model/stores';
-
 import { LoadingComponent, Modal } from '@/shared';
 import { PreviousButton } from '@/shared';
+import { useResponseFormStore, useGetFormResponseInfo, useHandleApplySubmit } from '@/entities/form';
+import { ExpiredForm, NoPermissionForm, ResponseForm } from '@/widgets/form/ui';
 
 const ApplyPage = () => {
   const params = useParams();
