+++ src/entities/form/index.ts
@@ -1,12 +1,5 @@
 export { Question } from './ui';
 export { SemesterMetaData } from './config';
-export {
-  EnrolledToggle,
-  CouncilFeeToggle,
-  SemesterToggle,
-  FormTitleInput,
-  LeaveOfAbsenceToggle,
-  GraduationToggle,
-  SemesterToggleForLeaveOfAbsence,
-} from './ui';
+export * from './ui';
 export * from './config';
+export * from './model';
