+++ src/entities/form/model/index.ts
@@ -1,3 +1,4 @@
 export * from './queries';
 export * from './mutations';
 export * from './stores';
+export * from './hooks';
