+++ src/entities/form/ui/FormStatusToggleBtn.tsx
@@ -2,8 +2,7 @@
 
 import { useParams } from 'next/navigation';
 
-import { useUpdateFormClosedStatus } from '@/entities/form/model';
-import { useFormResultStore } from '@/entities/form/model/stores';
+import { useFormResultStore, useUpdateFormClosedStatus } from '@/entities/form';
 
 export const FormStatusToggleBtn = () => {
   const params = useParams();
