+++ src/widgets/form/ui/FormRespondentInfo.tsx
@@ -1,5 +1,4 @@
-import { FORM_CONSTANT } from '@/entities/form';
-import { useFormResultStore } from '@/entities/form/model/stores';
+import { useFormResultStore, FORM_CONSTANT } from '@/entities/form';
 
 interface FormRespondentInfoProps {
   responseUserInfos: Form.ReplyResponseDto[];
