+++ src/widgets/form/ui/FormSummaryResult.tsx
@@ -2,14 +2,12 @@
 
 import { useParams } from 'next/navigation';
 
-import { useGetFormSummaryResult } from '@/entities/form/model';
-import { useFormResultStore } from '@/entities/form/model/stores';
 import {
   FormSummaryResultQuestionHeader,
   ObjectiveQuestionSummaryResult,
   SubjectiveQuestionSummaryResult,
 } from '@/entities/form/ui';
-
+import { useGetFormSummaryResult, useFormResultStore } from '@/entities/form';
 import { LoadingComponent } from '@/shared';
 
 export const FormSummaryResult = () => {
